@@ -1,60 +1,72 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.2.11'
-gem 'mysql2', '0.3.11'
+gem 'rails', '3.2.14'
+gem 'mysql2', '0.3.13'
 
 gem "delayed_job", "~> 3.0.4"
 gem "delayed_job_active_record", "~> 0.3.3"
 gem "daemons", "~> 1.1.9"
-gem "foreman", "~> 0.60.2"
+gem "foreman", "~> 0.63.0"
+gem "rake", "~> 10.1.0"
+gem "whenever", "~> 0.8.2", :require => false
 
-gem "libxml-ruby", "~> 2.4.0", :require => 'xml'
-gem "nokogiri", "~> 1.5.6"
-gem "multi_json", "~> 1.5.0"
-gem "oj", "~> 2.0.1"
-gem "nori", "~> 2.0.0"
+gem "faraday", "~> 0.8.7"
+gem "faraday_middleware", "~> 0.9.0"
+gem 'faraday-cookie_jar'
+gem "libxml-ruby", "~> 2.7.0", :require => 'xml'
+gem "nokogiri", "~> 1.6.0"
+gem "multi_json", "~> 1.7.8"
+gem "oj", "~> 2.1.4"
+gem "nori", "~> 2.3.0"
 gem 'mumboe-soap4r', '1.5.8.5'
-gem "devise", "~> 2.1.2"
+gem "devise", "~> 3.0.1"
+gem "omniauth-github", "~> 1.1.0"
+gem "omniauth-persona"
+gem "cancan", "~> 1.6.10"
 gem "validates_timeliness", "~> 3.0.14"
-gem "strip_attributes", "~> 1.3.0"
-gem "draper", "~> 1.1.0"
-gem "rabl", "~> 0.7.6"
+gem "strip_attributes", "~> 1.2"
+gem "draper", "~> 1.2.1"
+gem "rabl", "~> 0.8.6"
+gem "dalli", "~> 2.6.4"
+gem "cache_digests", "~> 0.3.1"
 
-gem "d3_rails", "~> 3.0.2"
-gem "bootstrap-sass", "~> 2.2.2.0"
-gem "turbolinks", "~> 1.0.0"
-gem 'will_paginate', '3.0.3'
+gem "d3_rails", "~> 3.2.4"
+gem "crossfilter-rails", "~> 1.2.0"
+gem "bootstrap-sass", "~> 2.3.2.1"
+gem "turbolinks", "~> 1.3.0"
+gem 'will_paginate', '3.0.4'
 gem "will_paginate-bootstrap", "~> 0.2.1"
-gem "simple_form", "~> 2.0.4"
+gem "simple_form", "~> 2.1.0"
 gem "dotiw", "~> 1.1.1"
-gem "seedbank", "~> 0.2.0"
+gem "github-markdown", "~> 0.5.3"
 
 group :assets do
-  gem "uglifier", "~> 1.3.0"
-  gem "jquery-rails", "~> 2.1.3"
+  gem "uglifier", "~> 2.1.2"
+  gem "jquery-rails", "~> 3.0.4"
   gem "therubyracer", "~> 0.11.1", :require => "v8"
   gem 'sass-rails', '~> 3.2.5'
   gem 'coffee-rails', '~> 3.2.2'
 end
 
 group :test do
-  gem "factory_girl_rails", "~> 4.1.0", :require => false
-  gem "cucumber-rails", "~> 1.3.0", :require => false
-  gem "capybara", "~> 1.1.4"
-  gem "database_cleaner", "~> 0.8.0"
+  gem "factory_girl_rails", "~> 4.2.1", :require => false
+  gem "cucumber-rails", "~> 1.3.1", :require => false
+  gem "capybara", "~> 2.1.0"
+  gem "database_cleaner", "~> 1.1.1"
   gem "launchy", "~> 2.1.2"
-  gem "email_spec", "~> 1.2.1"
+  gem "email_spec", "~> 1.5.0"
   gem "rack-test", "~> 0.6.2", :require => "rack/test"
   gem "simplecov", "~> 0.7.1", :require => false
-  gem "shoulda-matchers", "~> 1.4.1", :require => false
-  gem "webmock", "~> 1.9.0"
+  gem "shoulda-matchers", "~> 2.2.0", :require => false
+  gem "webmock", "~> 1.13.0"
   gem "rest-client", "~> 1.6.7"
-  gem "aruba", "~> 0.5.1"
-  gem "thin", "~> 1.5.0"
-  gem "poltergeist", "~> 1.0.2"
+  gem "aruba", "~> 0.5.1", :require => false
+  gem "poltergeist", "~> 1.3.0"
   gem "minitest", "~> 4.4.0"
+  gem "faye-websocket", "~> 0.4.7"
+  gem "mocha", "~> 0.13.3", :require => false
 end
 
 group :test, :development do
-  gem "rspec-rails", "~> 2.12.0"
+  gem "rspec-rails", "~> 2.14.0"
 end
