@@ -1,16 +1,21 @@
 Alm::Application.routes.draw do
-  devise_for :users, :controllers => { :registrations => "users/registrations" }
-
-  root :to => "index#index"
+  devise_for :users, :controllers => { :omniauth_callbacks => "users/omniauth_callbacks", :registrations => "users/registrations" }
+  
+  root :to => "docs#show"
 
   # constraints is added to allow dot in the url (doi is used to show article)
   resources :articles, :constraints => { :id => /.+?/, :format => /html|json|xml|csv/}
+  
   resources :sources
   resources :groups
+  resources :users
+  resources :docs, :only => :show, :constraints => { :id => /[0-z\-\.\(\)]+/ }
+  
+  match "oembed" => "oembed#show"
   
   namespace :admin do
     root :to => "index#index"
-    resources :articles, :constraints => { :id => /.+?/, :format => /html|js/}
+    resources :articles, :constraints => { :id => /.+?/, :format => /html|js/ }
     resources :sources
     resources :groups
     resources :delayed_jobs
@@ -19,12 +24,14 @@ Alm::Application.routes.draw do
     resources :responses
     resources :error_messages
     resources :api_requests
+    resources :users
   end
   
   namespace :api do
     namespace :v3 do
       root :to => "articles#index"
-      resources :articles, :constraints => { :id => /.+?/, :format=> false }
+      resources :articles, :constraints => { :id => /.+?/, :format=> false }, only: [:index, :show]
+      resources :articles, :constraints => { :id => /.+?/, :ip => /127.0.0.1/, :format=> false }
     end
   end
 
