@@ -1,8 +1,10 @@
 class Admin::DelayedJobsController < Admin::ApplicationController
-  
+
+  load_and_authorize_resource
+
   def index
     @sources = Source.active
-    respond_with @sources
+    render :index
   end
-  
+
 end
