@@ -1,15 +1,21 @@
 class Admin::EventsController < Admin::ApplicationController
   
+  skip_authorization_check
+  
   def index
-    sources = Source.active
-    articles = Source.active.map { |source| source.retrieval_statuses.count(:conditions => "event_count > 0") }
-    events = RetrievalStatus.joins(:source).where("sources.active = 1").order("group_id, display_name").group(:source_id).sum(:event_count).values
-    @data = sources.zip(articles, events).map { |source| { "name" => source.first.display_name, 
-                                                           "url" => admin_source_path(source.first),
-                                                           "group" => source.first.group_id, 
-                                                           "article_count" => source[1],
-                                                           "event_count" => source[2] } }.to_json
-    @article_count = Article.count
+    respond_with do |format|
+      format.html
+      format.json do
+        articles = Source.for_events.map { |source| source.retrieval_statuses.count(:conditions => "event_count > 0") }
+        events = RetrievalStatus.joins(:source).where("sources.active = 1").order("group_id, display_name").group(:source_id).sum(:event_count).values
+        @sources = Source.for_events.zip(articles, events).map { |source| { "name" => source.first.display_name, 
+                                                               "url" => admin_source_path(source.first),
+                                                               "group" => source.first.group_id, 
+                                                               "article_count" => source[1],
+                                                               "event_count" => source[2] } }
+        render :json => @sources
+      end
+    end
   end
   
 end
