@@ -22,27 +22,29 @@ class RetrievalStatus < ActiveRecord::Base
   include SourceHelper
 
   belongs_to :article, :touch => true
-  belongs_to :source
+  belongs_to :source, :touch => true
   has_many :retrieval_histories, :dependent => :destroy
-  
+
   serialize :event_metrics
-    
+
   delegate :name, :to => :source
+  delegate :display_name, :to => :source
 
   scope :most_cited, lambda { where("event_count > 0").order("event_count desc").limit(25) }
   scope :most_cited_last_x_days, lambda { |days| joins(:article).where("event_count > 0 AND articles.published_on >= CURDATE() - INTERVAL ? DAY", days).order("event_count desc").limit(25) }
   scope :most_cited_last_x_months, lambda { |months| joins(:article).where("event_count > 0 AND articles.published_on >= CURDATE() - INTERVAL ? MONTH", months).order("event_count desc").limit(25) }
-    
+
   scope :queued, where("queued_at is NOT NULL")
   scope :stale, where("queued_at is NULL AND scheduled_at IS NOT NULL AND scheduled_at <= NOW()")
   scope :published, joins(:article).where("queued_at is NULL AND articles.published_on <= CURDATE()")
-  
+  scope :with_sources, joins(:source).where("sources.active = 1").order("group_id, display_name")
+
   scope :total, lambda { |days| where("retrieved_at > NOW() - INTERVAL ? DAY", days) }
   scope :with_events, lambda { |days| where("event_count > 0 AND retrieved_at > NOW() - INTERVAL ? DAY", days) }
   scope :without_events, lambda { |days| where("event_count = 0 AND retrieved_at > NOW() - INTERVAL ? DAY", days) }
-  
+
   scope :by_source, lambda { |source_ids| where(:source_id => source_ids) }
-  
+
   def data
     if event_count > 0
       data = get_alm_data("#{source.name}:#{CGI.escape(article.doi)}")
@@ -50,18 +52,18 @@ class RetrievalStatus < ActiveRecord::Base
       nil
     end
   end
-  
+
   def events
-    unless data.blank? 
-      data["events"] 
+    unless data.blank?
+      data["events"]
     else
       []
     end
   end
-  
+
   def metrics
-    unless data.blank? 
-      data["event_metrics"] 
+    unless data.blank?
+      data["event_metrics"]
     else
       []
     end
@@ -193,7 +195,7 @@ class RetrievalStatus < ActiveRecord::Base
 
     end
   end
-  
+
   def delete_document
     unless data_rev.nil
       document_id = "#{source.name}:#{CGI.escape(article.uid)}"
@@ -202,7 +204,7 @@ class RetrievalStatus < ActiveRecord::Base
       nil
     end
   end
-  
+
   # calculate datetime when retrieval_status should be updated, adding random interval
   def stale_at
     unless article.published_on.nil?
@@ -223,9 +225,9 @@ class RetrievalStatus < ActiveRecord::Base
       random_time(source.staleness.last)
     end
   end
-  
+
   def random_time(duration)
     Time.zone.now + duration + rand(duration/10)
   end
-  
+
 end
