@@ -19,5 +19,6 @@
 Time::DATE_FORMATS.merge!(
     :friendly => '%m/%d/%y %I:%M%P UTC',
     :friendly_date => '%m/%d/%y',
-    :ordered => '%Y%m%d%H%M'
+    :ordered => '%Y%m%d%H%M',
+    :crossfilter => '%m%d%H%M'
 )
