@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Alm::Application.config.secret_token = 'cb695daf68dde14ed18e859eba68820feaf1f72469aa749b57642b300a9dbde32d5a5c55726969ed7ca5f01349ed403a169ffa44fe84ab1754cccbf4402e8a27'
+Alm::Application.config.secret_token = APP_CONFIG['secret_token']
