@@ -11,7 +11,7 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130211120421) do
+ActiveRecord::Schema.define(:version => 20130503125145) do
 
   create_table "api_requests", :force => true do |t|
     t.text     "path"
@@ -137,24 +137,29 @@ ActiveRecord::Schema.define(:version => 20130211120421) do
   add_index "sources", ["type"], :name => "index_sources_on_type", :unique => true
 
   create_table "users", :force => true do |t|
-    t.string   "email",                  :default => "", :null => false
-    t.string   "encrypted_password",     :default => "", :null => false
-    t.string   "reset_password_token"
-    t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
     t.integer  "sign_in_count",          :default => 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
-    t.string   "password_salt"
-    t.datetime "created_at",                             :null => false
-    t.datetime "updated_at",                             :null => false
+    t.datetime "created_at",                                 :null => false
+    t.datetime "updated_at",                                 :null => false
     t.string   "username"
+    t.string   "email",                  :default => "",     :null => false
+    t.string   "encrypted_password",     :default => "",     :null => false
+    t.string   "reset_password_token"
+    t.datetime "reset_password_sent_at"
+    t.string   "password_salt"
+    t.string   "provider"
+    t.string   "uid"
+    t.string   "name"
+    t.string   "authentication_token"
+    t.string   "role",                   :default => "user"
   end
 
-  add_index "users", ["email"], :name => "index_users_on_email", :unique => true
-  add_index "users", ["reset_password_token"], :name => "index_users_on_reset_password_token", :unique => true
+  add_index "users", ["authentication_token"], :name => "index_users_authentication_token", :unique => true
+  add_index "users", ["username"], :name => "index_users_username", :unique => true
 
   create_table "workers", :force => true do |t|
     t.integer  "identifier", :null => false
