@@ -1,7 +1,7 @@
 require 'source_helper'
 
 ### GIVEN ###
-Given /^that we have (\d+) API requests$/ do |number|
+Given /^we have (\d+) API requests$/ do |number|
   FactoryGirl.create_list(:api_request, number.to_i)
 end
 
@@ -11,4 +11,9 @@ When /^I make (\d+) API requests$/ do |number|
   (number.to_i).times do
     visit api_v3_article_path(article)
   end
+end
+
+### THEN ###
+Then(/^I should see (\d+) API requests were made$/) do |number|
+  page.has_css?('span#total', :text => number.to_s.reverse.gsub(/(\d{3})(?=\d)/, '\\1,').reverse).should be_true
 end
