@@ -10,5 +10,9 @@ end
 ### THEN ###
 Then /^I should see that CouchDB has (\d+) documents$/ do |number|
   page.driver.render("tmp/capybara/CouchDB.png")
-  page.has_css?('p#couchdb', :text => "#{number}CouchDB document")
+  page.has_css?('p#couchdb', :text => "#{number} CouchDB document").should be_true
+end
+
+Then(/^I should see the message "(.*?)" disappear$/) do |message|
+  page.has_no_css?('div#loading').should be_true
 end
