@@ -1,59 +1,47 @@
 ### UTILITY METHODS ###
-
-def create_visitor
-  @visitor ||= { :username => "example", :email => "example@example.com",
-    :password => "please", :password_confirmation => "please" }
+def create_user
+  @user = FactoryGirl.create(:user)
 end
 
 def find_user
-  @user ||= User.first conditions: {:email => @visitor[:email]}
-end
-
-def create_unconfirmed_user
-  create_visitor
-  delete_user
-  sign_up
-  visit '/users/sign_out'
-end
-
-def create_user
-  create_visitor
-  delete_user
-  @user = FactoryGirl.create(:user, email: @visitor[:email])
+  @user = FactoryGirl.create(:user)
 end
 
 def delete_user
-  @user ||= User.first conditions: {:email => @visitor[:email]}
   @user.destroy unless @user.nil?
 end
 
 def sign_up
   delete_user
-  visit '/users/sign_up'
-  fill_in "Username", :with => @visitor[:username]
-  fill_in "Email", :with => @visitor[:email]
-  fill_in "user_password", :with => @visitor[:password]
-  fill_in "user_password_confirmation", :with => @visitor[:password_confirmation]
-  click_button "Sign Up"
+  visit '/users/auth/github'
   find_user
 end
 
 def sign_in
-  visit '/users/sign_in'
-  fill_in "Username or email", :with => @visitor[:email]
-  fill_in "Password", :with => @visitor[:password]
-  click_button "Sign In"
+  visit '/users/auth/github'
   find_user
 end
 
 ### GIVEN ###
+Given /^we have (\d+) users$/  do |number|
+  FactoryGirl.create_list(:user, number.to_i - 1)
+end
+
+Given /^we have user "(.*?)" with name "(.*?)"$/ do |username, name|
+  FactoryGirl.create(:user, :username => username, :name => name)
+end
+
 Given /^I am not logged in$/ do
   visit '/users/sign_out'
 end
 
 Given /^I am logged in$/ do
-  create_user
-  sign_in
+  step "I am logged in as \"user\""
+end
+
+Given /^I am logged in as "(.*?)"$/ do |role|
+  FactoryGirl.create(:user, :role => role)
+  visit '/users/auth/github'
 end
 
 Given /^I exist as a user$/ do
@@ -61,17 +49,11 @@ Given /^I exist as a user$/ do
 end
 
 Given /^I do not exist as a user$/ do
-  create_visitor
   delete_user
 end
 
-Given /^I exist as an unconfirmed user$/ do
-  create_unconfirmed_user
-end
-
 ### WHEN ###
-When /^I sign in with valid credentials$/ do
-  create_visitor
+When /^I sign in$/ do
   sign_in
 end
 
@@ -79,103 +61,76 @@ When /^I sign out$/ do
   visit '/users/sign_out'
 end
 
-When /^I sign up$/ do
-  visit '/users/sign_up'
-end
-
-When /^I sign up with valid user data$/ do
-  create_visitor
-  sign_up
-end
-
-When /^I sign up with an invalid email$/ do
-  create_visitor
-  @visitor = @visitor.merge(:email => "notanemail")
-  sign_up
-end
-
-When /^I sign up without a password confirmation$/ do
-  create_visitor
-  @visitor = @visitor.merge(:password_confirmation => "")
-  sign_up
-end
-
-When /^I sign up without a password$/ do
-  create_visitor
-  @visitor = @visitor.merge(:password => "")
-  sign_up
-end
-
-When /^I sign up with a mismatched password confirmation$/ do
-  create_visitor
-  @visitor = @visitor.merge(:password_confirmation => "please123")
-  sign_up
-end
-
 When /^I return to the site$/ do
   visit '/'
 end
 
-When /^I sign in with a wrong email$/ do
-  @visitor = @visitor.merge(:email => "wrong@example.com")
-  sign_in
+When /^I go to my account page$/ do
+  visit '/users/me'
 end
 
-When /^I sign in with a wrong password$/ do
-  @visitor = @visitor.merge(:password => "wrongpass")
-  sign_in
+When /^I click on user "(.*?)"$/ do |username|
+  page.driver.render("tmp/capybara/#{username}.png")
+  user = User.find_by_username(username)
+  click_link "link_#{user.id}"
+end
+
+When /^I click on the Delete button for user "(.*?)" and confirm$/ do |username|
+  user = User.find_by_username(username)
+  within("#user_#{user.id}") do
+    click_link "#{user.id}-delete"
+  end
+end
+
+When /^I click on the "(.*?)" submenu of button Edit for user "(.*?)"$/ do |menu_item, username|
+  role = menu_item.split.last.downcase
+  user = User.find_by_username(username)
+  within("#user_#{user.id}") do
+    click_link "#{user.id}-update"
+    click_link "#{user.id}-update-#{role}"
+  end
 end
 
 ### THEN ###
-Then /^I should be signed in$/ do
-  page.should have_content "Sign Out"
-  page.should_not have_content "Sign Up"
-  page.should_not have_content "Sign In"
+Then /^I should see (\d+) user[s]?$/ do |number|
+  page.should have_css('div.accordion-group', :visible => true, :count => number.to_i)
 end
 
-Then /^I should not be signed up$/ do
-  page.should have_content "Invalid username"
-  page.should_not have_content "Sign Out"
+Then /^I should see user "(.*?)"$/ do |username|
+  user = User.find_by_username(username)
+  page.should have_css("a#link_#{user.id}")
+end
+
+Then /^I should not see user "(.*?)"$/ do |username|
+  user = User.find_by_username(username)
+  page.should_not have_css("a#link_#{user.id}")
+end
+
+Then /^I should be signed in$/ do
+  page.should have_css('#sign_out')
+  page.should_not have_css('#sign_in')
 end
 
 Then /^I should be signed out$/ do
-  page.should have_content "Sign In"
-  page.should_not have_content "Sign Out"
-end
-
-
-Then /^I see a successful sign in message$/ do
-  page.should have_content "Signed in successfully."
-end
-
-Then /^I should see a successful sign up message$/ do
-  page.should have_content "Welcome! You have signed up successfully."
-end
-
-Then /^I should see an invalid email message$/ do
-  page.should have_content "is invalid"
-end
-
-Then /^I should see a missing password message$/ do
-  page.should have_content "can't be blank"
-end
-
-Then /^I should see a missing password confirmation message$/ do
-  page.should have_content "doesn't match confirmation"
-end
-
-Then /^I should see a mismatched password message$/ do
-  page.should have_content "doesn't match confirmation"
-end
-
-Then /^I should see a signed out message$/ do
-  page.should have_content "Signed out successfully."
-end
-
-Then /^I see an invalid login message$/ do
-  page.should have_content "Invalid username / email or password."
+  page.should have_css('#sign_in')
+  page.should_not have_css('#sign_out')
 end
 
 Then /^I should reach the Sign In page$/ do
-  page.should have_content "Sign In"
+  page.should have_css('#sign_in')
+end
+
+Then /^I should not see the "(.*?)" button$/ do |title|
+  page.should_not have_link(title)
+end
+
+Then(/^I should see the API key$/) do
+  page.should have_css('dt', :text => "API Key")
+end
+
+Then /^I should see the "(.*?)" role for user "(.*?)"$/ do |role, username|
+  user = User.find_by_username(username)
+  within("#user_#{user.id}") do
+    page.should have_content role
+  end
 end
