@@ -17,6 +17,9 @@
 # limitations under the License.
 
 require 'net/http'
+require 'faraday'
+require 'faraday_middleware'
+require 'faraday-cookie_jar'
 
 module SourceHelper
   # default timeout is 60 sec
@@ -57,20 +60,10 @@ module SourceHelper
     end
 
     response = put_alm_data("#{service_url}#{id}", ActiveSupport::JSON.encode(data))
-    if response.kind_of?(Net::HTTPConflict)
-      # something went wrong
-      ErrorMessage.create(:exception => e, :message => "Failed to put #{service_url}#{id}. Going to try to get the document to get the current _rev, #{e.message}")   
-      
-      # get the most current revision value and use that to put the data one more time
-      cur_data = get_json("#{service_url}#{id}")
-      data[:_id] = cur_data["_id"]
-      data[:_rev] = cur_data["_rev"]
-
-      response = put_alm_data("#{service_url}#{id}", ActiveSupport::JSON.encode(data))
-    end
+    
+    return nil if response.nil?
 
     result = ActiveSupport::JSON.decode(response.body)
-
     result["rev"]
   end
 
@@ -78,30 +71,22 @@ module SourceHelper
     service_url = APP_CONFIG['couchdb_url']
     data = get_json("#{service_url}#{id}")
   end
-
-  def get_original_url(uri_str, limit = 10)
-    raise ArgumentError, 'too many HTTP redirects' if limit == 0
-
-    response = Net::HTTP.get_response(URI(uri_str))
-
-    case response
-    when Net::HTTPSuccess
-      uri_str
-    when Net::HTTPRedirection
-      location = response['location']
-
-      # sometimes we can get a location that doesn't have the host information
-      uri = URI(location)
-      if uri.host.nil?
-        orig_uri = URI(uri_str)
-        location = "http://" + orig_uri.host + location
-      end
-
-      get_original_url(location, limit - 1)
+  
+  def get_original_url(doi, limit = 10)
+    conn = Faraday.new(:url => "http://dx.doi.org") do |faraday|
+      faraday.use FaradayMiddleware::FollowRedirects, :limit => limit
+      faraday.use :cookie_jar
+      faraday.adapter Faraday.default_adapter
+    end
+    
+    response = conn.head doi
+    # Some publishers respond with a 403 error for the original URL
+    if response.status == 200 or (401..403) === response.status
+      response.env[:url].to_s
     else
       ErrorMessage.create(:exception => "", :class_name => response.class.to_s,
-                          :message => "Could not get the full url for #{uri_str}. #{response.message}, #{response.body}", 
-                          :status => response.code)
+                          :message => "Could not get the full URL for #{doi}, received #{response.env[:url].to_s} as last URL.", 
+                          :status => response.status)
       return ""
     end
   end
@@ -179,8 +164,8 @@ module SourceHelper
       Rails.logger.debug "[#{key}] = '#{value}']"
     end
 
-    # Store error_message and return empty body unless response is 2xx, 3xx or 404, don't raise an error
-    if [Net::HTTPSuccess, Net::HTTPOK, Net::HTTPRedirection, Net::HTTPNotFound].include?(response.class)
+    # Store error_message and return empty body unless response is 2xx or 404, don't raise an error
+    if response.kind_of?(Net::HTTPSuccess) or response.kind_of?(Net::HTTPNotFound)
       return response.body
     else
       ErrorMessage.create(:exception => "", :class_name => response.class.to_s,
@@ -263,10 +248,14 @@ module SourceHelper
     service_url = APP_CONFIG['couchdb_url']
     url = URI.parse(service_url)
     
-    res = Net::HTTP.start(url.host, url.port) { |http|http.request(req) }
-    unless res.kind_of?(Net::HTTPSuccess) or res.kind_of?(Net::HTTPNotFound)
-      res.error!
+    response = Net::HTTP.start(url.host, url.port) { |http|http.request(req) }
+    if response.kind_of?(Net::HTTPSuccess) or response.kind_of?(Net::HTTPNotFound)
+      response
+    else
+      ErrorMessage.create(:exception => "", :class_name => response.class.to_s,
+                          :message => "#{response.message} while requesting \"#{url.scheme}://#{url.host}:#{url.port}#{req.path}\"", 
+                          :status => response.code)
+      nil
     end
-    res
   end
 end
