@@ -3,12 +3,12 @@ require "spec_helper"
 describe "/api/v3/articles" do
   
   context "index" do
-    let(:articles) { FactoryGirl.create_list(:article, 55) }
+    let(:articles) { FactoryGirl.create_list(:article_with_events, 55) }
     
     context "more than 50 articles in query" do
       before(:each) do
         article_list = articles.collect { |article| "#{CGI.escape(article.doi)}" }.join(",") 
-        @uri = "/api/v3/articles?ids=#{article_list}&type=doi"
+        @uri = "/api/v3/articles?ids=#{article_list}&type=doi&api_key=12345"
       end
       
       it "JSON" do
@@ -27,7 +27,7 @@ describe "/api/v3/articles" do
         get @uri, nil, { 'HTTP_ACCEPT' => "application/xml" }
         last_response.status.should eql(200)
         
-        response = Nori.new.parse(last_response.body)
+        response = Nori.new(:advanced_typecasting => false).parse(last_response.body)
         response = response["articles"]["article"]
         response.length.should eql(50)
         response.any? do |article|
@@ -42,7 +42,7 @@ describe "/api/v3/articles" do
     
     context "show summary information" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?info=summary"}
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?info=summary&api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
@@ -58,26 +58,27 @@ describe "/api/v3/articles" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/xml" }
         last_response.status.should eql(200)
 
-        response_article = Nokogiri::XML(last_response.body).at_css("article")
-        response_article.content.should include(article.doi)
-        response_article.content.should include(article.published_on.to_time.utc.iso8601)
-        response_article.content.should_not include(article.sources.first.name)
+        response = Nori.new(:advanced_typecasting => false).parse(last_response.body)
+        response = response["articles"]["article"]
+        response["doi"].should eql(article.doi)
+        response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
+        response["sources"].should be_nil
       end
     end
   
-    context "historical data after 110 days" do
+    context "historical data after 30 days" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?days=110"}
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?days=30&api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
         last_response.status.should eql(200)
 
-        response_article = JSON.parse(last_response.body)[0]
-        response_source = response_article["sources"][0]
-        response_article["doi"].should eql(article.doi)
-        response_article["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
-        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.after_days(110).first.event_count)
+        response = JSON.parse(last_response.body)[0]
+        response_source = response["sources"][0]
+        response["doi"].should eql(article.doi)
+        response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
+        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.after_days(30).last.event_count)
         response_source["events"].should be_nil
         response_source["histories"].should be_nil
       end
@@ -91,26 +92,26 @@ describe "/api/v3/articles" do
         response_source = response["sources"]["source"]
         response["doi"].should eql(article.doi)
         response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
-        response_source["metrics"]["total"].to_i.should eq(article.retrieval_statuses.first.retrieval_histories.after_days(110).first.event_count)
+        response_source["metrics"]["total"].to_i.should eq(article.retrieval_statuses.first.retrieval_histories.after_days(30).last.event_count)
         response_source["events"].should be_nil
         response_source["histories"].should be_nil
       end
     
     end
   
-    context "historical data after 4 months" do
+    context "historical data after 6 months" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?months=4"}
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?months=6&api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
         last_response.status.should eql(200)
 
-        response_article = JSON.parse(last_response.body)[0]
-        response_source = response_article["sources"][0]
-        response_article["doi"].should eql(article.doi)
-        response_article["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
-        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.after_months(4).first.event_count)
+        response = JSON.parse(last_response.body)[0]
+        response_source = response["sources"][0]
+        response["doi"].should eql(article.doi)
+        response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
+        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.after_months(6).last.event_count)
         response_source["events"].should be_nil
         response_source["histories"].should be_nil
       end
@@ -124,7 +125,7 @@ describe "/api/v3/articles" do
         response_source = response["sources"]["source"]
         response["doi"].should eql(article.doi)
         response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
-        response_source["metrics"]["total"].to_i.should eq(article.retrieval_statuses.first.retrieval_histories.after_months(4).first.event_count)
+        response_source["metrics"]["total"].to_i.should eq(article.retrieval_statuses.first.retrieval_histories.after_months(6).last.event_count)
         response_source["events"].should be_nil
         response_source["histories"].should be_nil
       end
@@ -132,17 +133,17 @@ describe "/api/v3/articles" do
     
     context "historical data until 2012" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?year=2013"}
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?year=2012&api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
         last_response.status.should eql(200)
 
-        response_article = JSON.parse(last_response.body)[0]
-        response_source = response_article["sources"][0]
-        response_article["doi"].should eql(article.doi)
-        response_article["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
-        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.until_year(2013).first.event_count)
+        response = JSON.parse(last_response.body)[0]
+        response_source = response["sources"][0]
+        response["doi"].should eql(article.doi)
+        response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
+        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.until_year(2012).last.event_count)
         response_source["events"].should be_nil
         response_source["histories"].should be_nil
       end
@@ -156,7 +157,7 @@ describe "/api/v3/articles" do
         response_source = response["sources"]["source"]
         response["doi"].should eql(article.doi)
         response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
-        response_source["metrics"]["total"].to_i.should eq(article.retrieval_statuses.first.retrieval_histories.until_year(2013).first.event_count)
+        response_source["metrics"]["total"].to_i.should eq(article.retrieval_statuses.first.retrieval_histories.until_year(2012).last.event_count)
         response_source["events"].should be_nil
         response_source["histories"].should be_nil
       end
@@ -164,7 +165,7 @@ describe "/api/v3/articles" do
     
     context "show detail information" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?info=detail"}
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?info=detail&api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
@@ -199,7 +200,7 @@ describe "/api/v3/articles" do
     
     context "show history information" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?info=history"}
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?info=history&api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
@@ -232,7 +233,7 @@ describe "/api/v3/articles" do
     
     context "show event information" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?info=event"}
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?info=event&api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
