@@ -1,15 +1,43 @@
 require "spec_helper"
 
 describe "/api/v3/articles" do
-  let(:error) { { "error" => "No article found."} }
+  let(:error) { { "error" => "Article not found."} }
+  
+  context "missing api_key" do
+    let(:article) { FactoryGirl.create(:article_with_events) }
+    let(:uri) { "/api/v3/articles/info:doi/#{article.doi}"}
+    let(:missing_key) { { "error" => "Missing API key."} }
+
+    it "JSON" do
+      get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
+      last_response.status.should eql(401)
+      last_response.body.should eq(missing_key.to_json)
+      ErrorMessage.count.should == 1
+      error_message = ErrorMessage.first
+      error_message.class_name.should eq("Net::HTTPUnauthorized")
+      error_message.message.should include("Missing API key.")
+      error_message.status.should == 401
+    end
+  
+    it "XML" do
+      get uri, nil, { 'HTTP_ACCEPT' => "application/xml" }
+      last_response.status.should eql(401)
+      last_response.body.should eq(missing_key.to_xml)
+      ErrorMessage.count.should == 1
+      error_message = ErrorMessage.first
+      error_message.class_name.should eq("Net::HTTPUnauthorized")
+      error_message.message.should include("Missing API key.")
+      error_message.status.should == 401
+    end  
+  end
   
   context "index" do
-    let(:articles) { FactoryGirl.create_list(:article, 50) }
+    let(:articles) { FactoryGirl.create_list(:article_with_events, 50) }
     
     context "articles found via DOI" do
       before(:each) do
         article_list = articles.collect { |article| "#{CGI.escape(article.doi)}" }.join(",") 
-        @uri = "/api/v3/articles?ids=#{article_list}&type=doi"
+        @uri = "/api/v3/articles?ids=#{article_list}&type=doi&api_key=12345"
       end
     
       it "no format" do
@@ -40,7 +68,7 @@ describe "/api/v3/articles" do
         get @uri, nil, { 'HTTP_ACCEPT' => "application/xml" }
         last_response.status.should eql(200)
         
-        response = Nori.new.parse(last_response.body)
+        response = Nori.new(:advanced_typecasting => false).parse(last_response.body)
         response = response["articles"]["article"]
         response.length.should eql(50)
         response.any? do |article|
@@ -53,7 +81,7 @@ describe "/api/v3/articles" do
     context "articles found via PMID" do
       before(:each) do
         article_list = articles.collect { |article| "#{CGI.escape(article.pub_med)}" }.join(",") 
-        @uri = "/api/v3/articles?ids=#{article_list}&type=pmid"
+        @uri = "/api/v3/articles?ids=#{article_list}&type=pmid&api_key=12345"
       end
 
     
@@ -72,7 +100,7 @@ describe "/api/v3/articles" do
         get @uri, nil, { 'HTTP_ACCEPT' => "application/xml" }
         last_response.status.should eql(200)
         
-        response = Nori.new.parse(last_response.body)
+        response = Nori.new(:advanced_typecasting => false).parse(last_response.body)
         response = response["articles"]["article"]
         response.length.should eql(50)
         response.any? do |article|
@@ -83,7 +111,7 @@ describe "/api/v3/articles" do
     end
     
     context "no records found" do
-      let(:uri) { "/api/v3/articles"}
+      let(:uri) { "/api/v3/articles?api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
@@ -104,7 +132,7 @@ describe "/api/v3/articles" do
   
     context "DOI" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}"}
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}?api_key=12345"}
 
       it "no format" do
         get uri
@@ -114,7 +142,7 @@ describe "/api/v3/articles" do
         response_source = response_article["sources"][0]
         response_article["doi"].should eql(article.doi)
         response_article["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
-        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.last.event_count)
+        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.first.event_count)
         response_source["events"].should be_nil
         response_source["histories"].should be_nil
       end
@@ -127,7 +155,7 @@ describe "/api/v3/articles" do
         response_source = response["sources"][0]
         response["doi"].should eql(article.doi)
         response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
-        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.last.event_count)
+        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.first.event_count)
         response_source["events"].should be_nil
         response_source["histories"].should be_nil
       end
@@ -141,7 +169,7 @@ describe "/api/v3/articles" do
         response_source = response["sources"]["source"]
         response["doi"].should_not be_nil
         response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
-        response_source["metrics"]["total"].should eq(article.retrieval_statuses.first.retrieval_histories.last.event_count.to_s)
+        response_source["metrics"]["total"].to_i.should eq(article.retrieval_statuses.first.retrieval_histories.first.event_count)
         response_source["events"].should be_nil
         response_source["histories"].should be_nil
       end
@@ -150,7 +178,7 @@ describe "/api/v3/articles" do
   
     context "PMID" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:pmid/#{article.pub_med}"}
+      let(:uri) { "/api/v3/articles/info:pmid/#{article.pub_med}?api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
@@ -175,7 +203,7 @@ describe "/api/v3/articles" do
   
     context "PMCID" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:pmcid/PMC#{article.pub_med_central}"}
+      let(:uri) { "/api/v3/articles/info:pmcid/PMC#{article.pub_med_central}?api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
@@ -190,7 +218,7 @@ describe "/api/v3/articles" do
         last_response.status.should eql(200)
         
         response = Nori.new(:advanced_typecasting => false).parse(last_response.body)
-        response = response["articles"]["article"][0]
+        response = response["articles"]["article"]
         response["doi"].should_not be_nil
         response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
         response["pmcid"].should eql(article.pub_med_central.to_s)
@@ -200,7 +228,7 @@ describe "/api/v3/articles" do
   
     context "Mendeley" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:mendeley/#{article.mendeley}"}
+      let(:uri) { "/api/v3/articles/info:mendeley/#{article.mendeley}?api_key=12345"}
 
       it "JSON" do
         get uri, nil, { 'HTTP_ACCEPT' => "application/json" }
@@ -215,7 +243,7 @@ describe "/api/v3/articles" do
         last_response.status.should eql(200)
 
         response = Nori.new(:advanced_typecasting => false).parse(last_response.body)
-        response = response["articles"]["article"][0]
+        response = response["articles"]["article"]
         response["doi"].should_not be_nil
         response["publication_date"].should eql(article.published_on.to_time.utc.iso8601)
         response["mendeley"].should eql(article.mendeley)
@@ -225,7 +253,7 @@ describe "/api/v3/articles" do
       
     context "wrong DOI" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}xx"}     
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}xx?api_key=12345"}     
       
 
       it "JSON" do
@@ -243,7 +271,7 @@ describe "/api/v3/articles" do
     
     context "article not found when using format as file extension" do
       let(:article) { FactoryGirl.create(:article_with_events) }
-      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}xx"}      
+      let(:uri) { "/api/v3/articles/info:doi/#{article.doi}xx?api_key=12345"}      
 
       it "JSON" do
         get "#{uri}.json", nil, { 'HTTP_ACCEPT' => "application/json" }
