@@ -1,70 +1,80 @@
 FactoryGirl.define do
-  
+
   factory :article do
     sequence(:doi) {|n| "10.1371/journal.pone.00000#{n}" }
     sequence(:pub_med) {|n| "1897483#{n}" }
     pub_med_central 2568856
     mendeley "46cb51a0-6d08-11df-afb8-0026b95d30b2"
     title 'Defrosting the Digital Library: Bibliographic Tools for the Next Generation Web'
-    published_on { Time.zone.today - 1.day }
-    
+    published_on { Time.zone.today - 1.year }
+
     trait(:cited) { doi '10.1371/journal.pone.0000001' }
     trait(:uncited) { doi '10.1371/journal.pone.0000002' }
     trait(:not_publisher) { doi '10.1007/s00248-010-9734-2' }
     trait(:missing_mendeley) { mendeley nil }
     trait(:unpublished) { published_on { Time.zone.today + 1.week } }
-    
+
     factory :article_with_events do
       retrieval_statuses { |article| [article.association(:retrieval_status)] }
     end
-    
+
+    factory :article_for_feed do
+      published_on { Time.zone.today - 1.day }
+      retrieval_statuses { |article| [article.association(:retrieval_status, retrieved_at: Time.zone.today - 1.day)] }
+    end
+
     factory :article_with_errors do
       retrieval_statuses { |article| [article.association(:retrieval_status, :with_errors)] }
     end
-    
+
+    factory :article_with_private_citations do
+      retrieval_statuses { |article| [article.association(:retrieval_status, :with_private)] }
+    end
+
     factory :article_with_crossref_citations do
       retrieval_statuses { |article| [article.association(:retrieval_status, :with_crossref)] }
     end
-      
+
     factory :article_with_pubmed_citations do
       retrieval_statuses { |article| [article.association(:retrieval_status, :with_pubmed)] }
     end
-    
+
     factory :article_with_nature_citations do
       retrieval_statuses { |article| [article.association(:retrieval_status, :with_pubmed)] }
     end
-    
+
     factory :article_with_researchblogging_citations do
       retrieval_statuses { |article| [article.association(:retrieval_status, :with_researchblogging)] }
     end
   end
-  
+
   factory :group do
     name 'Citations'
-    
+
     initialize_with { Group.find_or_create_by_name(name) }
   end
-  
+
   factory :retrieval_history do
-    sequence(:retrieved_at) { Time.zone.today - 20.hours }
+    retrieved_at { Time.zone.today - 1.month }
     event_count { retrieval_status.event_count }
     status { event_count > 0 ? "SUCCESS" : "ERROR" }
   end
-  
+
   factory :retrieval_status do
     event_count 50
-    retrieved_at { Time.zone.now }
-    
+    retrieved_at { Time.zone.now - 1.month }
+
     association :article
     association :source, factory: :citeulike
-    
+
     trait(:unpublished) { association :article, :unpublished, factory: :article }
     trait(:missing_mendeley) do
-      association :article, :missing_mendeley, factory: :article 
+      association :article, :missing_mendeley, factory: :article
       association :source, factory: :mendeley
     end
     trait(:staleness) { association :source, factory: :citeulike }
     trait(:with_errors) { event_count 0 }
+    trait(:with_private) { association :source, private: true }
     trait(:with_crossref) { association :source, factory: :cross_ref }
     trait(:with_mendeley) { association :source, factory: :mendeley }
     trait(:with_pubmed) { association :source, factory: :pub_med }
@@ -72,18 +82,23 @@ FactoryGirl.define do
     trait(:with_researchblogging) { association :source, factory: :researchblogging }
     trait(:with_scienceseeker) { association :source, factory: :scienceseeker }
     trait(:with_wikipedia) { association :source, factory: :wikipedia }
-    
+
     before(:create) do |retrieval_status|
-      FactoryGirl.create_list(:retrieval_history, 
-                              5, 
-                              retrieval_status: retrieval_status, 
-                              article: retrieval_status.article, 
+      FactoryGirl.create(:retrieval_history,
+                              retrieved_at: Time.zone.today - 1.year + 1.day,
+                              event_count: 10,
+                              retrieval_status: retrieval_status,
+                              article: retrieval_status.article,
+                              source: retrieval_status.source)
+      FactoryGirl.create(:retrieval_history,
+                              retrieval_status: retrieval_status,
+                              article: retrieval_status.article,
                               source: retrieval_status.source)
     end
-          
+
     initialize_with { RetrievalStatus.find_or_create_by_article_id_and_source_id(article.id, source.id) }
   end
-  
+
   factory :citeulike, aliases: [:source], class: Citeulike do
     type "Citeulike"
     name "citeulike"
@@ -92,10 +107,10 @@ FactoryGirl.define do
     url "http://www.citeulike.org/api/posts/for/doi/%{doi}"
 
     group
-    
+
     initialize_with { Citeulike.find_or_create_by_name(name) }
   end
-  
+
   factory :copernicus, class: Copernicus do
     type "Copernicus"
     name "copernicus"
@@ -106,7 +121,7 @@ FactoryGirl.define do
     password "EXAMPLE"
 
     group
-    
+
     initialize_with { Copernicus.find_or_create_by_name(name) }
   end
   
@@ -133,23 +148,23 @@ FactoryGirl.define do
     password "EXAMPLE"
 
     group
-    
+
     initialize_with { CrossRef.find_or_create_by_name(name) }
   end
-  
+
   factory :nature, class: Nature do
     type "Nature"
     name "nature"
     display_name "Nature"
     active true
-    url "http://api.nature.com/service/blogs/posts.json?api_key=%{api_key}&doi=%{doi}"
+    url "http://blogs.nature.com/posts.json?api_key=%{api_key}&doi=%{doi}"
     api_key "EXAMPLE"
 
     group
-    
+
     initialize_with { Nature.find_or_create_by_name(name) }
   end
-  
+
   factory :pub_med, class: PubMed do
     type "PubMed"
     name "pubmed"
@@ -158,10 +173,10 @@ FactoryGirl.define do
     url "http://www.pubmedcentral.nih.gov/utils/entrez2pmcciting.cgi?view=xml&id=%{pub_med}"
 
     group
-    
+
     initialize_with { PubMed.find_or_create_by_name(name) }
   end
-  
+
   factory :researchblogging, class: Researchblogging do
     type "Researchblogging"
     name "researchblogging"
@@ -172,10 +187,10 @@ FactoryGirl.define do
     password "EXAMPLE"
 
     group
-    
+
     initialize_with { Researchblogging.find_or_create_by_name(name) }
   end
-  
+
   factory :science_seeker, class: ScienceSeeker do
     type "ScienceSeeker"
     name "scienceseeker"
@@ -184,7 +199,7 @@ FactoryGirl.define do
     url "http://scienceseeker.org/search/default/?type=post&filter0=citation&modifier0=doi&value0=%{doi}"
 
     group
-    
+
     initialize_with { ScienceSeeker.find_or_create_by_name(name) }
   end
   
@@ -222,7 +237,7 @@ FactoryGirl.define do
     url "http://%{host}/w/api.php?action=query&list=search&format=json&srsearch=%{doi}&srnamespace=0&srwhat=text&srinfo=totalhits&srprop=timestamp&srlimit=1"
 
     group
-    
+
     initialize_with { Wikipedia.find_or_create_by_name(name) }
   end
   
@@ -248,12 +263,12 @@ FactoryGirl.define do
     url_with_title "http://api.mendeley.com/oapi/documents/search/title:%{title}/?items=10&consumer_key=%{api_key}"
     related_articles_url "http://api.mendeley.com/oapi/documents/related/%{id}?consumer_key=%{api_key}"
     api_key "EXAMPLE"
-    
+
     group
-    
+
     initialize_with { Mendeley.find_or_create_by_name(name) }
   end
-  
+
   factory :facebook, class: Facebook do
     type "Facebook"
     name "facebook"
@@ -263,30 +278,35 @@ FactoryGirl.define do
     access_token "EXAMPLE"
 
     group
-    
+
     initialize_with { Facebook.find_or_create_by_name(name) }
   end
- 
+
   factory :user do
-    username 'example_user'
-    email 'example@example.com'
-    password 'please'
-    password_confirmation { |u| u.password }
+    sequence(:username) {|n| "joesmith#{n}" }
+    sequence(:name) {|n| "Joe Smith#{n}" }
+    sequence(:email) {|n| "joe#{n}@example.com" }
+    password "joesmith"
+    sequence(:authentication_token) {|n| "q9pWP8QxzkR24Mvs9BEy#{n}" }
+    role "admin"
+    provider "github"
+    uid "12345"
   end
-  
+
   factory :error_message do
     exception "An exception"
-    class_name "ActiveRecord::RecordNotFound"
-    message "Couldn't find Source with id=x"
+    class_name "Net::HTTPRequestTimeOut"
+    message "The request timed out."
     trace "backtrace"
     request "A request"
     user_agent "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/536.26.17 (KHTML, like Gecko) Version/6.0.2 Safari/536.26.17"
     target_url "http://127.0.0.1/sources/x"
-    status 404
+    status 408
     content_type "text/html"
   end
-  
+
   factory :api_request do
+    path "/api/v3/articles"
     page_duration 800
     db_duration 100
     view_duration 700
